Create AppUserMapper
utils contact with mapper and contacts with other mappers & passwordEncoder
